Create CNCS.ps1
The final ps1 incarnation of the main program. The Final Demo version will be an exe which will be in a zip folder. This is mostly just for test purposes as the file check was acting up and pulling a specific actions.ps1 file. So I want to rename it and have a final folder with the last versions of the files, thus here it is.